requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
openai==2.16.0
requests==2.32.5
langchain-core==1.2.7
langchain-community==0.4.1
langchain-text-splitters==1.1.0
qdrant-client==1.16.2
langchain-qdrant==1.1.0
fastembed==0.7.4
sentence-transformers==5.2.2
torch>=2.0,<3.0
pytest==8.3.3
streamlit==1.53.1
